move new client code to client package
